0329 ks 홈 등록폼쪽 쎼션작업

+++ ArtSpace/src/main/java/com/example/demo/hall/controller/HallController.java
@@ -15,11 +15,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.example.demo.SessionUtil;
 import com.example.demo.UUIDgeneration;
 import com.example.demo.hall.dto.EquipmentDTO;
 import com.example.demo.hall.dto.HallDTO;
 import com.example.demo.hall.service.HallService;
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 
 
@@ -29,15 +31,29 @@ public class HallController {
 
 	@Autowired
 	private HallService hallService;
-
+	
 	private Integer hall_id;
 	
+	SessionUtil user_session = new SessionUtil();
+	@Autowired
+	HttpSession session;
 	// 기본 공연장 정보 입력 화면 띄우기
 	@GetMapping("form")
 	public String showForm(Model model) {
-		HallDTO dto = new HallDTO();
-		model.addAttribute("hall_info", dto);
+		user_session.setSesstionValue(session);
+
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		} else if (!user_session.getAuthority().equals("SC")){
+			//법인 이용자인지 권한 체크
+			//return "redirect:/";
+		} else {
+			model.addAttribute("user_id", user_session.getUser_id());
+			model.addAttribute("nickname", user_session.getNickname());
+		}
 		
+		HallDTO dto = new HallDTO();
+		model.addAttribute("hall_info", dto);		
 		model.addAttribute("action", "/hall/form/insert");
 		return "html/hall/hall_form";
 	}
@@ -45,6 +61,15 @@ public class HallController {
 	// 이전 버튼을 눌렀을때 공연장 정보 수정 화면 띄우기
 	@GetMapping("form/{id}")
 	public String shwForm(Model model, @PathVariable("id") Integer id) {
+		user_session.setSesstionValue(session);
+		
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		} else {
+			model.addAttribute("user_id", user_session.getUser_id());
+			model.addAttribute("nickname", user_session.getNickname());
+		}
+
 		HallDTO hallInfo = hallService.findById(id);
 		hallInfo = hallService.findHallTime(hallInfo);
 		
@@ -54,15 +79,23 @@ public class HallController {
 		return "html/hall/hall_form";
 	}
 	
+	
 	// insert 처리
 	@PostMapping("/form/insert")
 	public String hallCreate(@ModelAttribute HallDTO hallDTO) {				
+		user_session.setSesstionValue(session);
+		
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		}
+
 		hall_id = hallService.findLastIndex();
 		if(hall_id == null) hall_id = 0;
 		hall_id++;
 		
 		hallDTO.setCreate_date(LocalDate.now());
 		hallDTO.setHall_id(hall_id);
+		hallDTO.setUser_id(user_session.getUser_id());
 
 		hallService.insert(hallDTO);
 		return "redirect:/hall/form/equipment/" + hallDTO.getHall_id();
@@ -72,8 +105,14 @@ public class HallController {
 	// update 처리
 	@PostMapping("/form/update/{id}")
 	public String hallUpdate(@ModelAttribute HallDTO hallDTO, @PathVariable("id") Integer id) {	
+		user_session.setSesstionValue(session);	
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		}
+
 		hallDTO.setCreate_date(LocalDate.now());
 		hallDTO.setHall_id(id);
+		hallDTO.setUser_id(user_session.getUser_id());
 		hallService.update(hallDTO);
 
 		return "redirect:/hall/form/equipment/" + id;
@@ -83,6 +122,15 @@ public class HallController {
 	// 장비 화면 띄우기
 	@GetMapping("form/equipment/{id}")
 	public String showFormEquipment(@PathVariable("id") Integer id, Model model) {
+		user_session.setSesstionValue(session);
+
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		} else {
+			model.addAttribute("user_id", user_session.getUser_id());
+			model.addAttribute("nickname", user_session.getNickname());
+		}
+
 		List<EquipmentDTO> equiList = hallService.getEquiList(id);
 		model.addAttribute("equiList", equiList);
 		
@@ -93,6 +141,11 @@ public class HallController {
 	// 이전 누르면 장비 저장하고 이전 화면으로 연결
 	@PostMapping("form/equipment/insert/{id}")
 	public String insertEquipment(@Valid @ModelAttribute HallDTO hallDTO, @PathVariable("id") Integer id) {
+		user_session.setSesstionValue(session);		
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		}
+
 		hallService.deleteEqui(id);	// 전부 삭제하고
 		List<EquipmentDTO> equiList = hallDTO.getEquiList();
 
@@ -109,6 +162,11 @@ public class HallController {
 	// 등록 누르면 장비 저장하고 my페이지로 연결
 	@PostMapping("form/complete/{id}")
 	public String completeInsertHall(@Valid @ModelAttribute HallDTO hallDTO, @PathVariable("id") Integer id) {
+		user_session.setSesstionValue(session);		
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		}
+
 		hallService.deleteEqui(id);	// 전부 삭제하고
 		List<EquipmentDTO> equiList = hallDTO.getEquiList();
 
@@ -124,6 +182,11 @@ public class HallController {
 	
 	@PostMapping("form/cancel/{id}")
 	public String cancelHall(@PathVariable("id") Integer id) {
+		user_session.setSesstionValue(session);		
+		if(user_session.getUser_id() == null) {
+			return "redirect:/login";
+		}
+
 		hallService.cancelHall(id);
 		return "redirect:/";
 	}
@@ -132,6 +195,12 @@ public class HallController {
 	// 임시 
 	@GetMapping("detail_page/{id}")
 	public String detailPage(Model model, HallDTO hallDTO, EquipmentDTO equipDTO, @PathVariable("id") Integer id) {
+		user_session.setSesstionValue(session);
+
+		if(user_session.getUser_id() != null) {
+			model.addAttribute("user_id", user_session.getUser_id());
+			model.addAttribute("nickname", user_session.getNickname());
+		}
 		
 		return "html/hall/detail_page";
 	}
